backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
python-dotenv>=1.0
pillow>=10.0
reportlab>=4.0
